tech-debt(backend): single-source ProviderStatus/ProvidersResponse media models (#12710)
ProviderStatus and ProvidersResponse were byte-identical duplicates in
api/image_generation.py and api/video_generation.py (only List vs list
typing differed). Moved to a new api/schemas_media.py module following
the existing schemas_*.py convention and imported in both routers.

No behavior change: response_model shape, JSON schema, and OpenAPI
component naming are unchanged for both /providers endpoints.
